requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
atlassian-python-api==4.0.7
attrs==25.3.0
Authlib==1.3.1
beautifulsoup4==4.13.5
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fastapi==0.101.1
frozenlist==1.7.0
grpcio
grpcio-health-checking
grpcio-tools
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.75
langchain-openai==0.3.29
langchain-text-splitters==0.3.11
langsmith==0.4.26
lxml==6.0.1
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.2
oauthlib==3.3.1
openai==1.93.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pdfminer.six==20221105
pdfplumber==0.9.0
pillow==11.3.0
propcache==0.3.2
protobuf
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.27.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
validators==0.34.0
Wand==0.6.13
weaviate-client==4.9.0
wrapt==1.17.3
xlsxwriter==3.2.5
yarl==1.20.1
zstandard==0.24.0
pyautogen
autogen